Convert SingularityMailer into an interface
